# improved parser performance