Added new algorithm to parse, passes all tests.
Added my own performance test based on disruptor performance tests
Tests results:

Fastest Dmitry
Starting performance tests
Run 0, Parse UUID=5,791 ops/sec
Run 1, Parse UUID=6,069 ops/sec
Run 2, Parse UUID=6,404 ops/sec
Run 3, Parse UUID=6,389 ops/sec
Run 4, Parse UUID=5,332 ops/sec
Run 5, Parse UUID=6,630 ops/sec
Run 6, Parse UUID=6,710 ops/sec
Run 7, Parse UUID=6,629 ops/sec
Run 8, Parse UUID=6,709 ops/sec
Run 9, Parse UUID=6,642 ops/sec
Fastest Noam
Starting performance tests
Run 0, Parse UUID=13,207 ops/sec
Run 1, Parse UUID=13,360 ops/sec
Run 2, Parse UUID=13,307 ops/sec
Run 3, Parse UUID=12,960 ops/sec
Run 4, Parse UUID=13,681 ops/sec
Run 5, Parse UUID=13,606 ops/sec
Run 6, Parse UUID=13,901 ops/sec
Run 7, Parse UUID=14,264 ops/sec
Run 8, Parse UUID=13,059 ops/sec
Run 9, Parse UUID=13,573 ops/sec