@dkomanov 2015-11-11
Take a look on an updated version (I've added yours). It's not the fastest one :P