@noam-almog 2015-11-11
+1
It looks like you added some new things and took the best from all worlds, looks very good.

I noticed the char translation can be optimized, my version actually made it slower, so great job. other than that, scala is faster ? interesting...

Take another look at the performance testing I added, it checks operation per sec but also it runs the methods in several runs over large number of iterations, larger than what you use. I think it's better to test it like that.